corrected the errors

+++ src/main/java/com/example/irctc/repository/TrainBodyRepository.java
@@ -3,5 +3,5 @@ package com.example.irctc.repository;
 import com.example.irctc.entity.TrainBody;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainBodyRepository extends JpaRepository<TrainBody,Integer> {
+public interface TrainBodyRepository extends JpaRepository<TrainBody,Long> {
 }
